requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
fdb==2.0.2
future==0.18.3
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
urllib3==2.0.4
